% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_anchor_cells.R
\name{adjustReferenceForPlatformEff}
\alias{adjustReferenceForPlatformEff}
\title{Platform effect adjustment on reference profiles based on the expression profiles of high confident anchors}
\usage{
adjustReferenceForPlatformEff(
  ref_profiles,
  obs_profiles,
  min_ref = 1,
  cutoff_ref = 0.6,
  cutoff_obs = 0.9,
  scaleFactor_limits = c(0.01, 100)
)
}
\arguments{
\item{ref_profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}

\item{obs_profiles}{Matrix of observed cluster mean profiles derived from high confident anchors in genes x cell types format.}

\item{min_ref}{Minimal expression values in reference profiles to define high expressors in reference, default = 1}

\item{cutoff_ref}{Cutoff in quantile to define high expressors of each cell type in reference profiles, default = 0.6}

\item{cutoff_obs}{Cutoff in quantile to define high expressors of each cell type in observed profiles, default = 0.9}

\item{scaleFactor_limits}{Vector of 2 elements to define the range of per gene scale factors to be consider in the adjusted reference profiles, set to NULL if not to exclude based on scale factor value, default = c(0.01, 100)}
}
\value{
A list with three elements: adj_profiles, Matrix of adjusted reference profiles in genes x cell types format; 
scaleFactor_DF, a data.frame with all genes of `ref_profiles` in rows and columns for `GeneName`, `scale_factor`, `type`; 
 and blacklist, a named vector with genes in name, scale factor in values, for genes with extreme scale factor beyond the range of `scaleFactor_limits` and thus not included in the returned `adj_profiles`, return NULL if no blacklist gene.
}
\description{
Calculates gene-wise scaling factor between reference profiles and the provided 
cluster mean profiles of high confidence anchors for high expressor genes in 
either profiles, and then adjusts the reference profiles accordingly to get 
the platform effect corrected profiles
}
