% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbclust.R
\name{Estep}
\alias{Estep}
\title{E step: estimate each cluster's mean profile}
\usage{
Estep(counts, clust, neg, assay_type)
}
\arguments{
\item{counts}{Counts matrix, cells * genes.}

\item{clust}{Vector of cluster assignments, or a matrix of probabilities
of cells (rows) belonging to clusters (columns).}

\item{neg}{Vector of mean background counts}

\item{assay_type}{Assay type of RNA, protein}
}
\value{
A matrix of cluster profiles, genes * clusters
}
\description{
Given cell assignments (or posterior probabilities), estimate the mean
 profile of each cluster.
}
