% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleProfiles.R
\name{estimatePlatformEffects}
\alias{estimatePlatformEffects}
\title{Platform effect adjustment on reference profiles based on the expression profiles of high confident anchors}
\usage{
estimatePlatformEffects(
  profiles,
  counts,
  neg,
  bg = NULL,
  anchors = NULL,
  min_ref = 1,
  cutoff_ref = 0.6,
  cutoff_obs = 0.9,
  scaleFactor_limits = c(0.01, 100)
)
}
\arguments{
\item{profiles}{Matrix of reference profiles holding mean expression of genes x cell types. 
Input linear-scale expression, with genes in rows and cell types in columns.}

\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{bg}{Expected background}

\item{anchors}{Vector giving "anchor" cell types, for use in semi-supervised
clustering. Vector elements will be mainly NA's (for non-anchored cells)
and cell type names for cells to be held constant throughout iterations.}

\item{min_ref}{Minimal expression values in reference profiles to define high expressors in reference, default = 1}

\item{cutoff_ref}{Cutoff in quantile to define high expressors of each cell type in reference profiles, default = 0.6}

\item{cutoff_obs}{Cutoff in quantile to define high expressors of each cell type in observed profiles, default = 0.9}

\item{scaleFactor_limits}{Vector of 2 elements to define the range of per gene scale factors to be consider in the adjusted reference profiles, set to NULL if not to exclude based on scale factor value, default = c(0.01, 100)}
}
\value{
A list with three elements: 
\describe{
    \item{adj_profiles}{Matrix of adjusted reference profiles in genes x cell types format}
    \item{scale_factor}{a named vector of gene-specific scaling factor for genes in `profiles`, with names in gene name}
    \item{blacklist}{a named vector with genes in name, scale factor in values, for genes with extreme scale factor beyond the range of `scaleFactor_limits` and thus not included in the returned `adj_profiles`, return NULL if no blacklist gene.}
}
}
\description{
Calculates gene-wise scaling factor between reference profiles and the provided 
cluster mean profiles of high confidence anchors for high expressor genes in 
either profiles, and then adjusts the reference profiles accordingly to get 
the platform effect corrected profiles
}
