% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lls_protein}
\alias{lls_protein}
\title{sum from negative binomial density function}
\usage{
lls_protein(mat, s, x, xsd, bg)
}
\arguments{
\item{mat}{dgCMatrix expression counts}

\item{s}{numeric scaling factor}

\item{x}{numeric expression for reference profile}

\item{xsd}{numeric value for reference SD}

\item{bg}{numeric background level}
}
\value{
rowSums for matrix of densities
}
\description{
Probability density function of the negative binomial distribution (written in C++)
}
