% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleProfiles.R
\name{updateReferenceProfiles}
\alias{updateReferenceProfiles}
\title{Update reference profiles}
\usage{
updateReferenceProfiles(
  reference_profiles,
  reference_sds,
  counts,
  neg,
  assay_type,
  bg = NULL,
  nb_size = 10,
  anchors = NULL,
  n_anchor_cells = 2000,
  min_anchor_cosine = 0.3,
  min_anchor_llr = 0.01,
  insufficient_anchors_thresh = 20,
  refinement = FALSE,
  blacklist = NULL,
  rescale = FALSE,
  refit = TRUE
)
}
\arguments{
\item{reference_profiles}{Matrix of reference mean profiles, genes * cell types}

\item{reference_sds}{Matrix of standard deviation profiles, genes * cell types. Only for assay_type of protein.}

\item{counts}{Counts matrix, cells * genes.}

\item{neg}{Vector of mean negprobe counts per cell}

\item{assay_type}{Assay type of RNA, protein}

\item{bg}{Expected background}

\item{nb_size}{The size parameter to assume for the NB distribution. Only for assay_type of RNA}

\item{anchors}{Vector giving "anchor" cell types, for use in semi-supervised
clustering. Vector elements will be mainly NA's (for non-anchored cells)
and cell type names for cells to be held constant throughout iterations.}

\item{n_anchor_cells}{For semi-supervised learning. Maximum number of anchor
cells to use for each cell type.}

\item{min_anchor_cosine}{For semi-supervised learning. Cells must have at
least this much cosine similarity to a fixed profile to be used as an
anchor.}

\item{min_anchor_llr}{For semi-supervised learning. Cells must have
(log-likelihood ratio / totalcounts) above this threshold to be used as an
anchor}
}
\value{
updated reference profiles
}
\description{
Update reference profiles using pre-specified anchor cells, or if no anchors
are specified, by first choosing anchor cells
}
